atlas_ops/requirements.txt
fastapi
uvicorn
sqlalchemy
pyyaml
docker
